+++ R/package-T4transport.R
@@ -6,7 +6,6 @@
 #' @importFrom Rdpack reprompt
 #' @import Rdpack
 #' @import CVXR
-#' @importFrom lpSolve lp
 #' @importFrom CVXR Variable Minimize matrix_trace Problem solve
 #' @importFrom stats rnorm median dnorm quantile cov
 #' @importFrom utils packageVersion