+++ src/utility_EMD.h
@@ -1,0 +1,11 @@
+#ifndef T4TRANSPORT_UTILITY_EMD_H
+#define T4TRANSPORT_UTILITY_EMD_H
+
+#define ARMA_NO_DEBUG
+#include <RcppArmadillo.h>
+
+arma::mat util_plan_emd_C(const arma::vec &a,
+                          const arma::vec &b,
+                          const arma::mat &C);
+
+#endif // T4TRANSPORT_UTILITY_EMD_H